+++ source/renderer/app/stores/NetworkStatusStore.js
@@ -481,8 +481,9 @@ export default class NetworkStatusStore extends Store {
   // DEFINE COMPUTED VALUES
   @computed get isIncentivizedTestnet(): boolean {
     return (
-      this.environment.isIncentivizedTestnet ||
-      this.nodeImplementation === NODE_IMPLEMENTATIONS.jormungandr
+      !this.environment.isTest &&
+      (this.environment.isIncentivizedTestnet ||
+        this.nodeImplementation === NODE_IMPLEMENTATIONS.jormungandr)
     );
   }