+++ storybook/stories/navigation/Sidebar.stories.js
@@ -11,6 +11,7 @@ import {
   STAGING,
 } from '../../../source/common/types/environment.types';
 import StoryDecorator from '../_support/StoryDecorator';
+import { isIncentivizedTestnetTheme } from '../_support/utils';
 import Sidebar from '../../../source/renderer/app/components/sidebar/Sidebar';
 import { WalletRecoveryPhraseVerificationStatuses } from '../../../source/renderer/app/stores/WalletsStore';
 import {
@@ -63,14 +64,11 @@ const sidebarMenus = observable({
 
 let emptyMenus;
 
-let currentTheme = sessionStorage.getItem('themeName') || 'light-blue';
-currentTheme = currentTheme.toLowerCase();
-
 storiesOf('Navigation|Sidebar', module)
   .addDecorator(story => <StoryDecorator>{story()}</StoryDecorator>)
   .addDecorator(withKnobs)
   // ====== Stories ======
-  .add('No Category', () => (
+  .add('No Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -81,12 +79,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
-      isIncentivizedTestnet
+      isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
     />
   ))
-  .add('Wallets Category', () => (
+  .add('Wallets Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -97,12 +95,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
-      isIncentivizedTestnet
+      isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
     />
   ))
-  .add('Wallet Selected', () => (
+  .add('Wallet Selected', (props: { currentTheme: string }) => (
     <Sidebar
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
       activeSidebarCategory={CATEGORIES_WITH_DELEGATION_COUNTDOWN[0].route}
@@ -114,12 +112,12 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
-      isIncentivizedTestnet
+      isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
     />
   ))
-  .add('Delegation Category', () => (
+  .add('Delegation Category', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -130,28 +128,31 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network="testnet"
-      isIncentivizedTestnet
+      isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
     />
   ))
-  .add('Decentralization Progress Category', () => (
-    <Sidebar
-      menus={emptyMenus}
-      categories={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN}
-      activeSidebarCategory={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN[1].route}
-      onActivateCategory={action('onActivateCategory')}
-      isDialogOpen={() => false}
-      onAddWallet={action('onAddWallet')}
-      onOpenDialog={action('openDialog')}
-      onSubmitSupportRequest={() => {}}
-      pathname="/"
-      currentTheme={currentTheme}
-      network="testnet"
-      isIncentivizedTestnet
-    />
-  ))
-  .add('Network label', () => (
+  .add(
+    'Decentralization Progress Category',
+    (props: { currentTheme: string }) => (
+      <Sidebar
+        menus={emptyMenus}
+        categories={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN}
+        activeSidebarCategory={CATEGORIES_WITHOUT_DELEGATION_COUNTDOWN[1].route}
+        onActivateCategory={action('onActivateCategory')}
+        isDialogOpen={() => false}
+        onAddWallet={action('onAddWallet')}
+        onOpenDialog={action('openDialog')}
+        onSubmitSupportRequest={() => {}}
+        pathname="/"
+        currentTheme={props.currentTheme}
+        network="testnet"
+        isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
+      />
+    )
+  )
+  .add('Network label', (props: { currentTheme: string }) => (
     <Sidebar
       menus={emptyMenus}
       categories={CATEGORIES_WITH_DELEGATION_COUNTDOWN}
@@ -162,7 +163,7 @@ storiesOf('Navigation|Sidebar', module)
       onOpenDialog={action('openDialog')}
       onSubmitSupportRequest={() => {}}
       pathname="/"
-      currentTheme={currentTheme}
+      currentTheme={props.currentTheme}
       network={select(
         'Netork badge',
         {
@@ -173,6 +174,6 @@ storiesOf('Navigation|Sidebar', module)
         },
         TESTNET
       )}
-      isIncentivizedTestnet
+      isIncentivizedTestnet={isIncentivizedTestnetTheme(props.currentTheme)}
     />
   ));